backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
aiofiles==23.2.1
rapidfuzz==3.9.4
pypdf==4.2.0
python-docx==1.1.2
python-dotenv==1.0.1
httpx==0.27.0
